added unit test for user reducer
